refactor(LineController): move responsibility of handling exceptions
- 코드의 관심사 분리를 좀 더 잘 지키기 위해 예외 처리를 다른 클래스로 위임
