+++ src/main/java/nextstep/subway/line/ui/LineController.java
@@ -77,39 +77,4 @@ public class LineController {
         lineService.deleteLine(lineId);
         return ResponseEntity.noContent().build();
     }
-
-    @ExceptionHandler(ConstraintViolationException.class)
-    public ResponseEntity handleConstraintViolationException(ConstraintViolationException e) {
-        return ResponseEntity.badRequest().build();
-    }
-
-    @ExceptionHandler(LineNotFoundException.class)
-    public ResponseEntity handleLineNotFoundException(LineNotFoundException e) {
-        return ResponseEntity.notFound().build();
-    }
-
-    @ExceptionHandler(InvalidSectionException.class)
-    public ResponseEntity handleInvalidSectionException(InvalidSectionException e) {
-        return ResponseEntity.badRequest().build();
-    }
-
-    @ExceptionHandler(StationNotFoundException.class)
-    public ResponseEntity handleStationNotFoundException(StationNotFoundException e) {
-        return ResponseEntity.badRequest().build();
-    }
-
-    @ExceptionHandler(TooLongSectionException.class)
-    public ResponseEntity handleTooLongSectionException(TooLongSectionException e) {
-        return ResponseEntity.badRequest().build();
-    }
-
-    @ExceptionHandler(TargetSectionNotFoundException.class)
-    public ResponseEntity handleTargetSectionNotFoundException(TargetSectionNotFoundException e) {
-        return ResponseEntity.badRequest().build();
-    }
-
-    @ExceptionHandler(EntityNotFoundException.class)
-    public ResponseEntity handleNotFoundException(EntityNotFoundException e) {
-        return ResponseEntity.notFound().build();
-    }
 }
